refactor: generateRandomCoins 로직 변경
- component -> domain 으로 이동하여 처리
